Move annotations to their own package
If I were new to the repository, would I know what "annotations" were
available to use with the model? Are they discoverable?

These questions led to moving these to a single package; the next step
would be documentation about their expected effects.
