+++ src/main/kotlin/model/CRCSystem.kt
@@ -3,6 +3,8 @@ package uk.gov.justice.hmpps.architecture
 import com.structurizr.model.Model
 import com.structurizr.model.SoftwareSystem
 import com.structurizr.view.ViewSet
+import uk.gov.justice.hmpps.architecture.annotations.ProblemArea
+import uk.gov.justice.hmpps.architecture.annotations.Tags
 
 class CRCSystem private constructor() {
   companion object : HMPPSSoftwareSystem {
