+++ src/main/kotlin/model/Delius.kt
@@ -6,6 +6,9 @@ import com.structurizr.model.Person
 import com.structurizr.model.SoftwareSystem
 import com.structurizr.view.AutomaticLayout
 import com.structurizr.view.ViewSet
+import uk.gov.justice.hmpps.architecture.annotations.APIDocs
+import uk.gov.justice.hmpps.architecture.annotations.ProblemArea
+import uk.gov.justice.hmpps.architecture.annotations.Tags
 
 class Delius private constructor() {
   companion object : HMPPSSoftwareSystem {
