+++ src/main/kotlin/model/EPF.kt
@@ -5,6 +5,7 @@ import com.structurizr.model.Person
 import com.structurizr.model.SoftwareSystem
 import com.structurizr.view.AutomaticLayout
 import com.structurizr.view.ViewSet
+import uk.gov.justice.hmpps.architecture.annotations.ProblemArea
 
 class EPF private constructor() {
   companion object : HMPPSSoftwareSystem {
