+++ src/main/kotlin/model/EQuiP.kt
@@ -3,6 +3,7 @@ package uk.gov.justice.hmpps.architecture
 import com.structurizr.model.Model
 import com.structurizr.model.SoftwareSystem
 import com.structurizr.view.ViewSet
+import uk.gov.justice.hmpps.architecture.annotations.ProblemArea
 
 class EQuiP private constructor() {
   companion object : HMPPSSoftwareSystem {
