+++ src/main/kotlin/model/IM.kt
@@ -5,6 +5,8 @@ import com.structurizr.model.Person
 import com.structurizr.model.SoftwareSystem
 import com.structurizr.view.AutomaticLayout
 import com.structurizr.view.ViewSet
+import uk.gov.justice.hmpps.architecture.annotations.ProblemArea
+import uk.gov.justice.hmpps.architecture.annotations.Tags
 
 class IM private constructor() {
   companion object : HMPPSSoftwareSystem {
