+++ src/main/kotlin/model/MoJSignOn.kt
@@ -4,6 +4,7 @@ import com.structurizr.model.Container
 import com.structurizr.model.Model
 import com.structurizr.model.SoftwareSystem
 import com.structurizr.view.ViewSet
+import uk.gov.justice.hmpps.architecture.annotations.Tags
 
 class MoJSignOn private constructor() {
   companion object : HMPPSSoftwareSystem {
