+++ src/main/kotlin/model/Pathfinder.kt
@@ -4,6 +4,7 @@ package uk.gov.justice.hmpps.architecture
 import com.structurizr.model.Container
 import com.structurizr.model.Model
 import com.structurizr.model.SoftwareSystem
+import uk.gov.justice.hmpps.architecture.annotations.Tags
 
 class Pathfinder(model: Model) {
   val system: SoftwareSystem
