+++ src/main/kotlin/model/PrisonVisitsBooking.kt
@@ -5,6 +5,7 @@ import com.structurizr.model.Model
 import com.structurizr.model.SoftwareSystem
 import com.structurizr.view.AutomaticLayout
 import com.structurizr.view.ViewSet
+import uk.gov.justice.hmpps.architecture.annotations.Tags
 
 class PrisonVisitsBooking private constructor() {
   companion object : HMPPSSoftwareSystem {
