+++ src/main/kotlin/model/ProbationTeamsService.kt
@@ -6,6 +6,8 @@ import com.structurizr.model.Model
 import com.structurizr.model.SoftwareSystem
 import com.structurizr.view.AutomaticLayout
 import com.structurizr.view.ViewSet
+import uk.gov.justice.hmpps.architecture.annotations.APIDocs
+import uk.gov.justice.hmpps.architecture.annotations.Tags
 
 class ProbationTeamsService private constructor() {
   companion object : HMPPSSoftwareSystem {
